+++ core/koin-core/src/commonMain/kotlin/org/koin/ext/StringExt.kt
@@ -15,4 +15,10 @@
  */
 package org.koin.ext
 
-fun String.quoted() = this.replace("\"", "")
+fun String.clearQuotes(): String {
+    val chars = this.toCharArray()
+    val quoteChar = '"'
+    return if (chars[0] == quoteChar && chars[chars.lastIndex] == quoteChar) {
+        chars.copyOfRange(1, chars.lastIndex).concatToString()
+    } else this
+}