+++ core/koin-core/src/jvmTest/kotlin/org/koin/core/StringTest.kt
@@ -1,0 +1,14 @@
+package org.koin.core
+
+import org.junit.Test
+import org.koin.ext.clearQuotes
+
+class StringTest {
+
+    @Test
+    fun quoted_strings(){
+        assert("test".clearQuotes() == "test")
+        assert("te\"st".clearQuotes() == "te\"st")
+        assert("\"test\"".clearQuotes() == "test")
+    }
+}